Fix Fixture, Model, add tests and yml file for test in pull request

+++ AirCompany.Domain/Models/AircraftModel.cs
@@ -23,7 +23,7 @@ public class AircraftModel
     /// <summary>
     /// Maximum flight range of the aircraft in kilometers.
     /// </summary>
-    public required int FlightRangeKm { get; set; }
+    public required double FlightRangeKm { get; set; }
 
     /// <summary>
     /// Maximum passenger capacity of the aircraft.
